Separate unrelated parts of StdLibrary into files
