+++ ulox/ulox.core/Package/Runtime/Library/StdLibrary.cs
@@ -1,29 +1,28 @@
 ﻿using System;
-using System.IO;
 
 namespace ULox
 {
     public class StdLibrary : IULoxLibrary
     {
         private const double SquareDividedTolerance = 0.01d;
-
+        
         public string Name => nameof(StdLibrary);
+
+        public Func<Vm> CreateVM { get; private set; }
 
         public StdLibrary()
         {
             CreateVM = () => new Vm();
         }
 
-        public Func<Vm> CreateVM { get; private set; }
-
         public Table GetBindings()
         {
             return this.GenerateBindingTable(
                 ("VM", Value.New(new VMClass(CreateVM))),
-                ("Factory", Value.New(MakeFactoryInstance())),
+                ("Factory", Value.New(FactoryStdLibrary.MakeFactoryInstance())),
                 ("Assert", Value.New(MakeAssertInstance())),
-                ("Serialise", Value.New(MakeSerialiseInstance())),
-                ("DI", Value.New(MakeDIInstance())),
+                ("Serialise", Value.New(SerialiseStdLibrary.MakeSerialiseInstance())),
+                ("DI", Value.New(DIStdLibrary.MakeDIInstance())),
                 (nameof(Duplicate), Value.New(Duplicate)),
                 (nameof(str), Value.New(str)),
                 (nameof(IsFrozen), Value.New(IsFrozen)),
@@ -34,50 +33,7 @@ namespace ULox
                                             );
         }
 
-        private InstanceInternal MakeFactoryInstance()
-        {
-            var factoryInst = new InstanceInternal();
-            factoryInst.AddFieldsToInstance(
-                (nameof(Line), Value.New(Line)),
-                (nameof(SetLine), Value.New(SetLine))
-                );
-            factoryInst.Freeze();
-            return factoryInst;
-        }
-
-        public NativeCallResult Line(Vm vm, int argCount)
-        {
-            var arg = vm.GetArg(1);
-            var line = vm.Factory.GetLine(arg);
-            vm.PushReturn(line);
-            return NativeCallResult.SuccessfulExpression;
-        }
-
-        public NativeCallResult SetLine(Vm vm, int argCount)
-        {
-            var key = vm.GetArg(1);
-            if (key.IsNull)
-                throw new VMException($"'{nameof(SetLine)}' must have non null key argument.");
-            var line = vm.GetArg(2);
-            if(line.IsNull)
-                throw new VMException($"'{nameof(SetLine)}' must have non null line argument.");
-            
-            vm.Factory.SetLine(key, line);
-            return NativeCallResult.SuccessfulExpression;
-        }
-
-        private InstanceInternal MakeDIInstance()
-        {
-            var diLibInst = new InstanceInternal();
-            diLibInst.AddFieldsToInstance(
-                (nameof(Count), Value.New(Count)),
-                (nameof(GenerateDump), Value.New(GenerateDump)),
-                (nameof(Freeze), Value.New(Freeze)));
-            diLibInst.Freeze();
-            return diLibInst;
-        }
-
-        private InstanceInternal MakeAssertInstance()
+        internal InstanceInternal MakeAssertInstance()
         {
             var assertInst = new InstanceInternal();
             assertInst.AddFieldsToInstance(
@@ -95,102 +51,6 @@ namespace ULox
                 (nameof(Fail), Value.New(Fail)));
             assertInst.Freeze();
             return assertInst;
-        }
-
-        private static InstanceInternal MakeSerialiseInstance()
-        {
-            var serialiseInst = new InstanceInternal();
-            serialiseInst.AddFieldsToInstance(
-                (nameof(ToJson), Value.New(ToJson)),
-                (nameof(FromJson), Value.New(FromJson)));
-            serialiseInst.Freeze();
-            return serialiseInst;
-        }
-
-        public NativeCallResult GenerateStackDump(Vm vm, int argCount)
-        {
-            vm.PushReturn(Value.New(vm.GenerateStackDump()));
-            return NativeCallResult.SuccessfulExpression;
-        }
-
-        public NativeCallResult GenerateGlobalsDump(Vm vm, int argCount)
-        {
-            vm.PushReturn(Value.New(vm.GenerateGlobalsDump()));
-            return NativeCallResult.SuccessfulExpression;
-        }
-
-        public NativeCallResult GenerateReturnDump(Vm vm, int argCount)
-        {
-            vm.PushReturn(Value.New(vm.GenerateReturnDump()));
-            return NativeCallResult.SuccessfulExpression;
-        }
-
-        public NativeCallResult IsFrozen(Vm vm, int argCount)
-        {
-            var target = vm.GetArg(1);
-            if (target.type == ValueType.Instance)
-                vm.PushReturn(Value.New(target.val.asInstance.IsFrozen));
-            else if (target.type == ValueType.Class)
-                vm.PushReturn(Value.New(target.val.asClass.IsFrozen));
-
-            return NativeCallResult.SuccessfulExpression;
-        }
-
-        public NativeCallResult Unfreeze(Vm vm, int argCount)
-        {
-            var target = vm.GetArg(1);
-            if (target.type == ValueType.Instance)
-                target.val.asInstance.Unfreeze();
-            if (target.type == ValueType.Class)
-                target.val.asClass.Unfreeze();
-
-            return NativeCallResult.SuccessfulExpression;
-        }
-
-        private NativeCallResult Count(Vm vm, int argCount)
-        {
-            var di = FromVm(vm);
-            vm.PushReturn(Value.New(di.Count));
-            return NativeCallResult.SuccessfulExpression;
-        }
-
-        private NativeCallResult GenerateDump(Vm vm, int argCount)
-        {
-            var di = FromVm(vm);
-            vm.PushReturn(Value.New(di.GenerateDump()));
-            return NativeCallResult.SuccessfulExpression;
-        }
-
-        private NativeCallResult Freeze(Vm vm, int argCount)
-        {
-            var di = FromVm(vm);
-            di.Freeze();
-            return NativeCallResult.SuccessfulExpression;
-        }
-
-        private DiContainer FromVm(Vm vMBase)
-            => vMBase.DiContainer;
-
-        private static NativeCallResult ToJson(Vm vm, int argCount)
-        {
-            var obj = vm.GetArg(1);
-            var jsonWriter = new JsonValueHeirarchyWriter();
-            var walker = new ValueHeirarchyWalker(jsonWriter);
-            walker.Walk(obj);
-            var result = jsonWriter.GetString();
-            vm.PushReturn(Value.New(result));
-            return NativeCallResult.SuccessfulExpression;
-        }
-
-        private static NativeCallResult FromJson(Vm vm, int argCount)
-        {
-            var jsonString = vm.GetArg(1);
-            var reader = new StringReader(jsonString.val.asString.String);
-            var creator = new JsonDocValueHeirarchyTraverser(new ValueObjectBuilder(ValueObjectBuilder.ObjectType.Object), reader);
-            creator.Process();
-            var obj = creator.Finish();
-            vm.PushReturn(obj);
-            return NativeCallResult.SuccessfulExpression;
         }
 
         private static NativeCallResult AreApproxEqual(Vm vm, int argCount)
@@ -322,6 +182,47 @@ namespace ULox
         {
             var msg = vm.GetArg(1);
             throw new AssertException($"Fail. '{msg}'");
+            return NativeCallResult.SuccessfulExpression;
+        }
+
+
+        public NativeCallResult GenerateStackDump(Vm vm, int argCount)
+        {
+            vm.PushReturn(Value.New(vm.GenerateStackDump()));
+            return NativeCallResult.SuccessfulExpression;
+        }
+
+        public NativeCallResult GenerateGlobalsDump(Vm vm, int argCount)
+        {
+            vm.PushReturn(Value.New(vm.GenerateGlobalsDump()));
+            return NativeCallResult.SuccessfulExpression;
+        }
+
+        public NativeCallResult GenerateReturnDump(Vm vm, int argCount)
+        {
+            vm.PushReturn(Value.New(vm.GenerateReturnDump()));
+            return NativeCallResult.SuccessfulExpression;
+        }
+
+        public NativeCallResult IsFrozen(Vm vm, int argCount)
+        {
+            var target = vm.GetArg(1);
+            if (target.type == ValueType.Instance)
+                vm.PushReturn(Value.New(target.val.asInstance.IsFrozen));
+            else if (target.type == ValueType.Class)
+                vm.PushReturn(Value.New(target.val.asClass.IsFrozen));
+
+            return NativeCallResult.SuccessfulExpression;
+        }
+
+        public NativeCallResult Unfreeze(Vm vm, int argCount)
+        {
+            var target = vm.GetArg(1);
+            if (target.type == ValueType.Instance)
+                target.val.asInstance.Unfreeze();
+            if (target.type == ValueType.Class)
+                target.val.asClass.Unfreeze();
+
             return NativeCallResult.SuccessfulExpression;
         }
